Move InMemoryStorage class to storage package
