+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportTest.java
@@ -34,6 +34,7 @@ import co.rsk.peg.feeperkb.FeePerKbResponseCode;
 import co.rsk.peg.feeperkb.FeePerKbStorageProvider;
 import co.rsk.peg.feeperkb.FeePerKbSupport;
 import co.rsk.peg.storage.BridgeStorageAccessorImpl;
+import co.rsk.peg.storage.InMemoryStorage;
 import co.rsk.peg.storage.StorageAccessor;
 import co.rsk.peg.bitcoin.*;
 import co.rsk.peg.btcLockSender.*;
